requirements.txt
flask
openai>=1.0.0
python-dotenv
gunicorn==21.2.0
